Debug

+++ tests/test1.c
@@ -1,2 +1,2 @@
-float raio = 10;
+int raio = 10;
 float area = 3.14 * raio * raio;
